Mover función extraerMensajesError a utils
